pyproject.toml
[project]
name = "pubmed"
version = "0.1.0"
description = ""
authors = [
    {name = "Amal Rajan",email = "amalrajan.in@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "biopython (>=1.85,<2.0)"
]

[tool.poetry]
packages = [{include = "pubmed", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
responses = "^0.25.7"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
get-papers-list = "pubmed.cli:main"
